# 008-crud-example-with-echo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple server to demonstrate the CRUD operations via the echo framework.",
        "title": "Simple CRUD example",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "paths": {
        "/users/{id}": {
            "get": {
                "description": "Gets the user by ID.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the user",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorView"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorView"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorView"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ErrorView": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "lastName": {
                    "type": "string"
                }
            }
        }
    }
}